+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASE
 
+- **Add** : Added Bancomat Pay payment method
 - **Add** : Added Apple Pay multi-browser configuration and checkout support with popup/modal display modes.
 
 ## 1.31.2